Move `MessageBatch` to `stateful`
