+++ packages/engine/src/datastore/batch/mod.rs
@@ -3,13 +3,11 @@ pub mod boolean;
 pub mod context;
 pub mod dataset;
 pub mod iterators;
-pub mod message;
 pub mod migration;
 
-pub use stateful::agent::AgentBatch;
+pub use stateful::{agent::AgentBatch, message::MessageBatch};
 
 pub use self::{
     context::{AgentIndex, ContextBatch},
     dataset::Dataset,
-    message::MessageBatch,
 };
